Feat/add readme

+++ database/migrations/2022_10_27_093154_create_notifications_table.php
@@ -15,7 +15,7 @@ return new class () extends Migration {
         Schema::create('notifications', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained();
-            $table->unsignedSmallInteger('user_to_notify')->constrained()->cascadeOnDelete();
+            $table->unsignedBigInteger('user_to_notify')->constrained()->cascadeOnDelete();
             $table->string('type');
             $table->boolean('seen_by_user')->nullable();
             $table->timestamps();
